fix(layout): rename CollapsePanel -> Sidebar
