+++ packages/layout/src/index.ts
@@ -3,4 +3,5 @@ export * from './layout'
 export * from './main'
 export * from './navigation'
 export * from './panel'
+export * from './sidebar'
 export * from './navbar'
