+++ packages/layout/src/navigation/navigation-header/NavigationHeader.tsx
@@ -2,16 +2,16 @@
 
 import { useContext } from 'react'
 import { clsx, Text, TextProps } from '@midas-ds/components'
-import { CollapsePanelContext } from '../../panel'
+import { SidebarContext } from '../../sidebar'
 import styles from './NavigationHeader.module.css'
 
 export const NavigationHeader = ({ className, ...rest }: TextProps) => {
-  const { isCollapsed } = useContext(CollapsePanelContext)
+  const sidebarContext = useContext(SidebarContext)
 
   return (
     <Text
       className={clsx(className, styles.navigationHeader, {
-        [styles.collapsed]: isCollapsed,
+        [styles.collapsed]: sidebarContext?.isCollapsed,
       })}
       {...rest}
     />
