+++ packages/layout/src/navigation/navigation-item/NavigationItem.tsx
@@ -2,7 +2,7 @@
 
 import clsx from 'clsx'
 import { DetailedHTMLProps, HTMLAttributes, useContext } from 'react'
-import { CollapsePanelContext } from '../../panel'
+import { SidebarContext } from '../../sidebar'
 import styles from './NavigationItem.module.css'
 
 export type NavigationItemProps = DetailedHTMLProps<
@@ -11,12 +11,12 @@ export type NavigationItemProps = DetailedHTMLProps<
 >
 
 export const NavigationItem = ({ className, ...rest }: NavigationItemProps) => {
-  const { isCollapsed } = useContext(CollapsePanelContext)
+  const sidebarContext = useContext(SidebarContext)
 
   return (
     <li
       className={clsx(className, styles.navigationItem, {
-        [styles.collapsed]: isCollapsed,
+        [styles.collapsed]: sidebarContext?.isCollapsed,
       })}
       {...rest}
     />
