+++ packages/layout/src/navigation/navigation-section/NavigationSection.tsx
@@ -3,10 +3,10 @@
 import clsx from 'clsx'
 import { useContext } from 'react'
 import { Collection, CollectionProps } from '@react-aria/collections'
-import { CollapsePanelContext } from '../../panel'
 import { NavigationHeader } from '../navigation-header'
-import styles from './NavigationSection.module.css'
 import { NavigationItem } from '../navigation-item'
+import { SidebarContext } from '../../sidebar'
+import styles from './NavigationSection.module.css'
 
 export interface NavigationSectionProps<
   T extends object,
@@ -20,13 +20,13 @@ export const NavigationSection = <T extends object>({
   title,
   ...rest
 }: NavigationSectionProps<T>) => {
-  const { isCollapsed } = useContext(CollapsePanelContext)
+  const sidebarContext = useContext(SidebarContext)
 
   return (
     <NavigationItem>
       <ul
         className={clsx(className, styles.navigationSection, {
-          [styles.collapsed]: isCollapsed,
+          [styles.collapsed]: sidebarContext?.isCollapsed,
         })}
       >
         {title && <NavigationHeader>{title}</NavigationHeader>}
