--- packages/layout/src/panel/collapse-panel/index.ts
@@ -1,2 +1,0 @@
-export * from './CollapsePanel'
-export * from './CollapsePanelContext'
