--- packages/layout/src/panel/dismiss-panel/useDismissPanels.ts
@@ -1,8 +1,0 @@
-'use client'
-
-import { useContext } from 'react'
-import { DismissPanelContext } from './DismissPanelContext'
-
-export const useDismissPanels = () => {
-  return useContext(DismissPanelContext)
-}
