+++ packages/layout/src/panel/index.ts
@@ -1,7 +1,8 @@
-export * from './Panel'
-export * from './PanelContext'
-export * from './collapse-panel'
-export * from './dismiss-panel'
 export * from './panel-body'
 export * from './panel-header'
 export * from './panel-title'
+export * from './Panel'
+export * from './PanelContext'
+export * from './PanelProvider'
+export * from './PanelRegion'
+export * from './usePanels'
