+++ packages/layout/src/panel/intl/translations.json
@@ -1,12 +1,8 @@
 {
   "en": {
-    "expandSidebar": "Expand sidebar",
-    "collapseSidebar": "Collapse sidebar",
     "closePanel": "Close panel"
   },
   "sv": {
-    "expandSidebar": "Expandera sidopanel",
-    "collapseSidebar": "Minimera sidopanel",
     "closePanel": "Stäng panel"
   }
 }
