+++ packages/layout/src/panel/usePanels.ts
@@ -1,0 +1,6 @@
+'use client'
+
+import { useContext } from 'react'
+import { PanelContext } from './PanelContext'
+
+export const usePanels = () => useContext(PanelContext)
